requirements.txt
dbt-core==1.7.10
astronomer-cosmos==1.3.2
# apache-airflow-providers-amazon==3.0.0
# apache-airflow-providers-apache-hdfs==3.2.1
# apache-airflow-providers-apache-hive==6.0.0
apache-airflow-providers-apache-spark==4.0.1
# apache-airflow-providers-apache-sqoop==3.1.1
# apache-airflow-providers-celery==3.1.0
# apache-airflow-providers-cncf-kubernetes==5.2.2
# apache-airflow-providers-common-sql==1.3.4
# apache-airflow-providers-docker==3.5.1
# apache-airflow-providers-elasticsearch==4.4.0
# apache-airflow-providers-ftp==3.3.1
# apache-airflow-providers-google==8.11.0
# apache-airflow-providers-grpc==3.1.0
# apache-airflow-providers-hashicorp==3.3.0
# apache-airflow-providers-http==4.2.0
# apache-airflow-providers-imap==3.1.1
# apache-airflow-providers-microsoft-azure==5.2.1
# apache-airflow-providers-mysql==4.0.2
# apache-airflow-providers-odbc==3.2.1
# apache-airflow-providers-postgres==5.4.0
# apache-airflow-providers-presto==5.0.0
# apache-airflow-providers-redis==3.1.0
# apache-airflow-providers-sendgrid==3.1.0
# apache-airflow-providers-sftp==4.2.4
# apache-airflow-providers-slack==7.2.0
# apache-airflow-providers-snowflake==4.0.4
# apache-airflow-providers-sqlite==3.3.1
# apache-airflow-providers-ssh==3.5.0
# airflow-mcd==0.1.2
# pycarlo==0.7.5
# amplitude-analytics==1.1.1
# PyMySQL
# salesforce-bulk
# mysql-connector-python
# ruamel.yaml==0.17.21
# ruamel.yaml.clib==0.2.6
